Made it compile and run again
I'm still running into segfaults though

+++ Rasteriser.cpp
@@ -246,8 +246,8 @@ void Rasteriser::drawLine(const LinePrimitive& line) const
 	const vec3 posB_ndc = vec3(posB_clip) / posB_clip.w;
 	const vec3 posB_win = viewport->calculateWindowCoordinates( posB_ndc );
 
-	ivec2 a = ivec2(posA_win);
-	ivec2 b = ivec2(posB_win);
+	vec2 a = vec2(posA_win);
+	vec2 b = vec2(posB_win);
 
 	float lineLength = length(b-a); 
 	unsigned int positionCounter = 0;
